+++ tests/HelperNumericTest.php
@@ -20,8 +20,11 @@ class HelperNumericTest extends HelperTestCase
         self::assertEquals('01', HelperNumeric::formatAmountDigits(1, 2));
 
         self::assertEquals(HelperNumeric::formatAmountDigits(1, 1), '1');
+
         self::assertEquals(HelperNumeric::formatAmountDigits(1, 2), '01');
+
         self::assertEquals(HelperNumeric::formatAmountDigits(1, 3), '001');
+
         self::assertEquals(HelperNumeric::formatAmountDigits(100, 1), '100');
     }
 
@@ -30,8 +33,15 @@ class HelperNumericTest extends HelperTestCase
      */
     public function testIntImplode(): void
     {
-        self::assertSame(HelperNumeric::intImplode([3, 5, 1, 3, 4, 2, 2], '-', false, true), '3-5-1-4-2');
-        self::assertSame(HelperNumeric::intImplode([3, 5, 1, 3, 4, 2.3, 2.7]), '1,2,2,3,4,5');
+        self::assertSame(
+            HelperNumeric::intImplode([3, 5, 1, 3, 4, 2, 2], '-', false, true),
+            '3-5-1-4-2'
+        );
+
+        self::assertSame(
+            HelperNumeric::intImplode([3, 5, 1, 3, 4, 2.3, 2.7]),
+            '1,2,2,3,4,5'
+        );
     }
 
     /**
@@ -39,11 +49,20 @@ class HelperNumericTest extends HelperTestCase
      */
     public function testIntExplode(): void
     {
-        self::assertSame(json_encode(HelperNumeric::intExplode('3-5-1-4-2-2', '-')), '[3,5,1,4,2,2]');
+        self::assertSame(
+            json_encode(HelperNumeric::intExplode('3-5-1-4-2-2', '-')),
+            '[3,5,1,4,2,2]'
+        );
 
-        //        self::assertSame(json_encode(HelperNumeric::intExplode(null)), '[]');
+//        self::assertSame(
+//          json_encode(HelperNumeric::intExplode(null)),
+//          '[]'
+//        );
 
-        self::assertSame(json_encode(HelperNumeric::intExplode('')), '[0]');
+        self::assertSame(
+            json_encode(HelperNumeric::intExplode('')),
+            '[0]'
+        );
     }
 
     /**
@@ -51,11 +70,20 @@ class HelperNumericTest extends HelperTestCase
      */
     public function testFloatExplode(): void
     {
-        self::assertSame(json_encode(HelperNumeric::floatExplode('3.3-5-1.5-4-2-2', '-')), '[3.3,5,1.5,4,2,2]');
+        self::assertSame(
+            json_encode(HelperNumeric::floatExplode('3.3-5-1.5-4-2-2', '-')),
+            '[3.3,5,1.5,4,2,2]'
+        );
 
-        self::assertSame(json_encode(HelperNumeric::floatExplode('3.3-5-1.5-4-2-2', '.')), '[3,3,5]');
+        self::assertSame(
+            json_encode(HelperNumeric::floatExplode('3.3-5-1.5-4-2-2', '.')),
+            '[3,3,5]'
+        );
 
-        //        self::assertSame(json_encode(HelperNumeric::floatExplode(null, '.', false)), '[]');
+//        self::assertSame(
+//          json_encode(HelperNumeric::floatExplode(null, '.', false)),
+//          '[]'
+//        );
     }
 
     /**
@@ -63,10 +91,19 @@ class HelperNumericTest extends HelperTestCase
      */
     public function testCalcBytesSize(): void
     {
-        self::assertSame(json_encode(HelperNumeric::calcBytesSize(234)), '{"size":234,"unit":"B"}');
+        self::assertSame(
+            json_encode(HelperNumeric::calcBytesSize(234)),
+            '{"size":234,"unit":"B"}'
+        );
 
-        self::assertSame(json_encode(HelperNumeric::calcBytesSize(1150)), '{"size":1.1,"unit":"KB"}');
+        self::assertSame(
+            json_encode(HelperNumeric::calcBytesSize(1150)),
+            '{"size":1.1,"unit":"KB"}'
+        );
 
-        self::assertSame(json_encode(HelperNumeric::calcBytesSize(4500000)), '{"size":4.4,"unit":"MB"}');
+        self::assertSame(
+            json_encode(HelperNumeric::calcBytesSize(4500000)),
+            '{"size":4.4,"unit":"MB"}'
+        );
     }
 }