+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.0.2 - Released 2020/10/14
 -----------------------------------
 * Resolve all known issues of static analysis
 * Add PHPStan static analysis to to travis/ci inspection
+* Bugfix HelperReflection::callUserFunction() after failed unit test
 
 Version 1.0.1 - Released 2020/10/13
 -----------------------------------